+++ image-map.tf
@@ -1,15 +1,15 @@
 locals {
         image_region_map = {
-          "hpcc-lsf10-scale5131-rhel84-2-0-5" = {
-            "us-south" = "r006-748b60db-c08c-4787-af8d-8fe826658ec9"
-            "us-east"  = "r014-592de712-4eb9-4cea-a818-4bf4463941c1"
-            "au-syd"   = "r026-f9f08627-8a6c-49c7-82a8-baa18a2a32a3"
-            "jp-osa"   = "r034-06996b45-4eaf-4e9c-81a1-821fe566cdf0"
-            "jp-tok"   = "r022-81cfdc61-348f-4f56-99f2-e0e6f639c8c5"
-            "eu-de"    = "r010-d69501d1-4358-48ff-b384-91eac2c60b28"
-            "eu-gb"    = "r018-66daa0bb-5b37-49c8-b703-d8a1cc92c3cf"
-            "ca-tor"   = "r038-9b5eff8f-2be8-4162-a95c-f1967d95ee19"
-            "br-sao"   = "r042-d11fc922-8649-4df9-a5a4-d3e27264448a"
+          "hpcc-lsf10-scale5151-rhel84-2-1" = {
+            "us-south" = "r006-81fd1256-bd98-4dc6-9a27-f43ad1db685f"
+            "us-east"  = "r014-bd78eb0f-bedc-432d-ba89-16ffb16ba1e6"
+            "au-syd"   = "r026-389d641d-fea8-4b86-8701-2894a09f315f"
+            "jp-osa"   = "r034-a18449a3-979b-41a0-9ebd-2831ccbbaaef"
+            "jp-tok"   = "r022-160e642c-2b9f-480f-bc04-2a06f9f200cb"
+            "eu-de"    = "r010-ee38025e-28ff-443f-b935-0afc84d3ea5e"
+            "eu-gb"    = "r018-0567f781-b141-4ee8-a767-af84bcce13b4"
+            "ca-tor"   = "r038-abbbaf63-6b82-487f-b96c-7ae06430265d"
+            "br-sao"   = "r042-aad163df-3049-4b6a-b5b1-735ff6ffec92"
           }
         }
 }