+++ resources/ibmcloud/security/security_tcp_rule.tf
@@ -18,10 +18,9 @@ variable "sg_direction" {}
 variable "remote_ip_addr" {}
 
 resource "ibm_is_security_group_rule" "itself" {
-  count = length(var.remote_ip_addr)
   group     = var.security_group_id
   direction = var.sg_direction
-  remote    = var.remote_ip_addr[count.index]
+  remote    = var.remote_ip_addr[0]
 
     tcp {
       port_min = 22
@@ -30,5 +29,5 @@ resource "ibm_is_security_group_rule" "itself" {
 }
 
 output "security_rule_id" {
-  value = ibm_is_security_group_rule.itself[*].rule_id
+  value = ibm_is_security_group_rule.itself.rule_id
 }