+++ scripts/LSF_management_static_candidate_server.sh
@@ -11,14 +11,14 @@ echo START `date '+%Y-%m-%d %H:%M:%S'` >> $logfile
 
 nfs_server=${storage_ips}
 nfs_mount_dir="data"
-controller_ips=($controller_ips)
-lsfcontrollerhost=${controller_ips[0]}
+management_host_ips=($management_host_ips)
+lsfmanagement_host=${management_host_ips[0]}
 #cluster_name=""
-#lsfcontrollerhost=""
+#lsfmanagement_host=""
 
 #default value for the host name prefix
 vmPrefix="icgen2host"
-lsfcontrollerhost=${vmPrefix}-${lsfcontrollerhost//./-}
+lsfmanagement_host=${vmPrefix}-${lsfmanagement_host//./-}
 
 if [ ! -z $cluster_name ]
 then
@@ -29,30 +29,22 @@ fi
 
 #If no dns, then will fixed the hostname based on provate IP address and hostname, if you have dns server, then can completely remove this part
 privateIP=$(ip addr show eth0 | awk '$1 == "inet" {gsub(/\/.*$/, "", $2); print $2}')
-ManagementCandidateHostName=${vmPrefix}-${privateIP//./-}
+Management_Host_Candidate_Name=${vmPrefix}-${privateIP//./-}
 networkIPrange=$(echo ${privateIP}|cut -f1-3 -d .)
-hostnamectl set-hostname ${ManagementCandidateHostName}
+hostnamectl set-hostname ${Management_Host_Candidate_Name}
 host_prefix=$(hostname|cut -f1-4 -d -)
 
-
 # Change the MTU setting as this is required for setting mtu as 9000 for communication to happen between clusters
-ip route replace $rc_cidr_block  dev eth0 proto kernel scope link src $privateIP mtu 9000
-echo 'ip route replace '$rc_cidr_block' dev eth0 proto kernel scope link src '$privateIP' mtu 9000' >> /etc/sysconfig/network-scripts/route-eth0
-
-#for controllerIP in $controller_ips; do
-  #if [ "$controllerIP" != "$privateIP" ]; then
-      #ip route add $controllerIP dev eth0 mtu 9000
-      #echo 'ip route add '$controllerIP' dev eth0 mtu 9000' >> /etc/sysconfig/network-scripts/route-eth0
-  #fi
-#done
+echo "MTU=9000" >> "/etc/sysconfig/network-scripts/ifcfg-eth0"
+systemctl restart NetworkManager
 
 
 # NOTE: On ibm gen2, the default DNS server do not have reverse hostname/IP resolution.
-# 1) put the controller server hostname and ip into lsf hosts.
+# 1) put the management_host server name and ip into lsf hosts.
 # 2) put all possible VMs' hostname and ip into lsf hosts.
 python3 -c "import ipaddress; print('\n'.join([str(ip) + ' ${vmPrefix}-' + str(ip).replace('.', '-') for ip in ipaddress.IPv4Network('${rc_cidr_block}')]))" >> /etc/hosts
 
-#Update Controller host name based on with nfs share or not
+#Update Management_host name based on with nfs share or not
 if ([ -n "${nfs_server}" ] && [ -n "${nfs_mount_dir}" ]); then
   echo "NFS server and share found, start mount nfs share!" >> $logfile
   #Mount the nfs share
@@ -63,8 +55,8 @@ if ([ -n "${nfs_server}" ] && [ -n "${nfs_mount_dir}" ]); then
   #make auto mount when server is down
   echo "$nfs_server:/$nfs_mount_dir /mnt/${nfs_mount_dir} nfs rw,rsize=1048576,wsize=1048576,hard,timeo=600,retrans=2,_netdev 0 0 " >> /etc/fstab
   echo "Mount nfs share done!" >> $logfile
-  while [ ! -d /mnt/$nfs_mount_dir/lsf_$lsfcontrollerhost ]; do sleep 1s; done
-  if [ -d /mnt/$nfs_mount_dir/lsf_$lsfcontrollerhost ]; then
+  while [ ! -d /mnt/$nfs_mount_dir/lsf_$lsfmanagement_host ]; do sleep 1s; done
+  if [ -d /mnt/$nfs_mount_dir/lsf_$lsfmanagement_host ]; then
     echo "lsf directory already exits in nfs share" >> $logfile
     for subdir in conf work log das_staging_area; do
       lsf_link=$(ls -la /opt/ibm/lsf/$subdir | grep "\->")
@@ -73,7 +65,7 @@ if ([ -n "${nfs_server}" ] && [ -n "${nfs_mount_dir}" ]); then
       else
         echo "link the conf to share location" >> $logfile
         mv /opt/ibm/lsf/${subdir} /opt/ibm/lsf/${subdir}_org
-        ln -fs /mnt/$nfs_mount_dir/lsf_$lsfcontrollerhost/$subdir /opt/ibm/lsf/$subdir
+        ln -fs /mnt/$nfs_mount_dir/lsf_$lsfmanagement_host/$subdir /opt/ibm/lsf/$subdir
       fi
     done
   else
@@ -87,7 +79,9 @@ if ([ -n "${nfs_server}" ] && [ -n "${nfs_mount_dir}" ]); then
   cat /mnt/$nfs_mount_dir/ssh/id_rsa.pub >> /root/.ssh/authorized_keys
   cp /mnt/$nfs_mount_dir/ssh/id_rsa /home/lsfadmin/.ssh/
   cp /mnt/$nfs_mount_dir/ssh/authorized_keys /home/lsfadmin/.ssh/authorized_keys
-  echo "${temp_public_key}" >> /root/.ssh/authorized_keys
+  if [ ${spectrum_scale} == true ]; then
+      echo "${temp_public_key}" >> /root/.ssh/authorized_keys
+  fi
   chmod 600 /home/lsfadmin/.ssh/authorized_keys
   chmod 700 /home/lsfadmin/.ssh
   chown -R lsfadmin:lsfadmin /home/lsfadmin/.ssh
@@ -146,7 +140,7 @@ EOF
 
 # TODO: disallow root login
 
-#startup lsf daemons in the management candidate nodes
+#startup lsf daemons in the management_host candidate nodes
 
 echo 1 > /proc/sys/vm/overcommit_memory # new image requires this. otherwise, it reports many failures of memory allocation at fork() if we use candidates. why?
 echo 'vm.overcommit_memory=1' > /etc/sysctl.d/90-lsf.conf