+++ scripts/LSF_spectrum_storage.sh
@@ -9,25 +9,23 @@
 LSF_TOP=/opt/ibm/lsf_worker
 LSF_CONF=$LSF_TOP/conf
 LSF_CONF_FILE=$LSF_CONF/lsf.conf
-LSF_ENTITLEMENT_FILE=$LSF_CONF/lsf.entitlement
-LS_ENTITLEMENT_FILE=$LSF_CONF/ls.entitlement
 LSF_HOSTS_FILE=$LSF_CONF/hosts
 . $LSF_TOP/conf/profile.lsf
 DATA_DIR=/data
 
 env
 
-#Update controller host name based on internal IP address
+#Update management_host name based on internal IP address
 vmPrefix="icgen2host"
 nfs_mount_dir="data"
 privateIP=$(ip addr show eth0 | awk '$1 == "inet" {gsub(/\/.*$/, "", $2); print $2}')
 hostName=${vmPrefix}-${privateIP//./-}
 networkIPrange=$(echo ${privateIP}|cut -f1-3 -d .)
 hostnamectl set-hostname ${hostName}
-controllerHostNamesStr=`echo "${controller_ips//./-}" | sed -e 's/^/ibm-gen2host-/g' | sed -e 's/ / ibm-gen2host-/g'`
+management_hostNamesStr=`echo "${management_host_ips//./-}" | sed -e 's/^/ibm-gen2host-/g' | sed -e 's/ / ibm-gen2host-/g'`
 
 # NOTE: On ibm gen2, the default DNS server do not have reverse hostname/IP resolution.
-# 1) put the controller server hostname and ip into lsf hosts.
+# 1) put the management_host server name and ip into lsf hosts.
 # 2) put all possible VMs' hostname and ip into lsf hosts.
 python3 -c "import ipaddress; print('\n'.join([str(ip) + ' ${vmPrefix}-' + str(ip).replace('.', '-') for ip in ipaddress.IPv4Network('${rc_cidr_block}')]))" >> /etc/hosts
 mkdir -p /mnt/$nfs_mount_dir
@@ -39,8 +37,8 @@ while [ ! -f /mnt/$nfs_mount_dir/ssh/id_rsa ]; do sleep 5; let retry++; echo Ret
 ln -s /mnt/$nfs_mount_dir /root/shared
 
 # Change the MTU setting as this is required for setting mtu as 9000 for communication to happen between clusters
-ip route replace $rc_cidr_block  dev eth0 proto kernel scope link src $privateIP mtu 9000
-echo 'ip route replace '$rc_cidr_block' dev eth0 proto kernel scope link src '$privateIP' mtu 9000' >> /etc/sysconfig/network-scripts/route-eth0
+echo "MTU=9000" >> "/etc/sysconfig/network-scripts/ifcfg-eth0"
+systemctl restart NetworkManager
 
 mkdir -p /root/.ssh
 cat /mnt/$nfs_mount_dir/ssh/id_rsa.pub >> /root/.ssh/authorized_keys